Completed work needed to add interaction radii to the model

+++ src/edu/mtu/simulation/decay/DecayModel.java
@@ -19,7 +19,7 @@ public interface DecayModel {
 	 * @param moleclues The current number of molecules.
 	 * @return The number of the molecules that should decay.
 	 */
-	public long getDecayQuantity(int timeStep, String compound, long moleclues);
+	public double getDecayQuantity(int timeStep, String compound, long moleclues);
 	
 	/**
 	 * Prepare the decay model by loading the experimental results,
